requirements.txt
mcp>=1.12,<2
httpx>=0.27,<1
python-dotenv>=1.0,<2
